🔧 Fix #70

+++ examples/tacotron2/decode_tacotron2.py
@@ -113,7 +113,7 @@ def main():
         mel_outputs, post_mel_outputs, stop_outputs, alignment_historys = tacotron2.inference(
             charactor,
             char_length,
-            speaker_ids=tf.zeros(shape=[tf.shape(charactor)[0]]),
+            speaker_ids=tf.zeros(shape=[tf.shape(charactor)[0]], dtype=tf.int32),
         )
 
         # convert to numpy
